# openapi.yaml
openapi: 3.0.3
info:
  title: APS Fire Alarm API
  version: 1.0.0
  description: API schema for authentication, devices, telemetry, and alerts.
paths:
  /alerts/:
    get:
      operationId: alerts_list
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      tags:
      - Alerts
      security:
      - cookieAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAlertList'
          description: ''
...
  /devices/{id}/alerts/:
    get:
      operationId: devices_alerts_retrieve
      summary: List alerts for a device
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: Device ID (integer)
        required: true
      - in: query
        name: status
        schema:
          type: string
        description: 'Filter by status: open|resolved'
      tags:
      - Alerts
      security:
      - cookieAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
          description: ''
  /devices/{id}/telemetry/:
    get:
      operationId: devices_telemetry_retrieve
      summary: List telemetry for a device
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: Device ID (integer)
        required: true
      - in: query
        name: since
        schema:
          type: string
        description: ISO8601 datetime or epoch seconds (>=)
      - in: query
        name: until
        schema:
          type: string
        description: ISO8601 datetime or epoch seconds (<=)
      tags:
      - Telemetry
      security:
      - cookieAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
          description: ''
  /devices/register/:
    post:
      operationId: devices_register_create
      description: |
        Register a device as either a master (default) or a slave.

        How to register a slave device:
        - Set `device_role` to `slave`.
        - Provide `master_id` referencing an existing master device that you own (non-admin users).
        - The selected master must have role `master`; you cannot attach to another slave.

        Additional rules:
        - `master_id` is required when `device_role` is `slave`, and must NOT be provided when `device_role` is `master`.
        - Latitude and longitude are required on first registration.
        - If the hardware identifier is already registered by another user, the endpoint returns 409 Conflict.
        - If the device is already registered by you (or you are superadmin), the same call updates name/coordinates and returns 200.
      summary: Register/claim a device
      tags:
      - Devices
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceRegister'
            examples:
              RegisterMasterRequest:
                value:
                  hardware_identifier: MASTER-001
                  device_name: Main Panel
                  latitude: 23.777628
                  longitude: 90.405449
                  device_role: master
              RegisterSlaveRequest:
                value:
                  hardware_identifier: SLAVE-101
                  device_name: Floor 1 Sensor
                  latitude: 23.7777
                  longitude: 90.4055
                  device_role: slave
                  master_id: 123
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/DeviceRegister'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DeviceRegister'
        required: true
      security:
      - cookieAuth: []
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
          description: ''
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
          description: ''
        '409':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
              examples:
                DeviceAlreadyRegistered:
                  value:
                    detail: Device already registered by another user
          description: Conflict when attempting to register a device owned by another
            user.
  /devices/tree/:
    get:
      operationId: devices_tree_retrieve
      summary: List devices as a tree (masters with nested slaves)
      tags:
      - Devices
      security:
      - cookieAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedDeviceTreeList'
              examples:
                DevicesTreeResponse:
                  value:
                    count: 123
                    next: http://api.example.org/accounts/?page=4
                    previous: http://api.example.org/accounts/?page=2
                    results:
                    - - id: 123
                        hardware_identifier: MASTER-001
                        device_name: Main Panel
                        device_role: master
                        slaves:
                        - id: 456
                          hardware_identifier: SLAVE-101
                          device_role: slave
                          master_id: 123
                        - id: 789
                          hardware_identifier: SLAVE-102
                          device_role: slave
                          master_id: 123
          description: ''
...
  /telemetry/:
    get:
      operationId: telemetry_list
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      tags:
      - Telemetry
      security:
      - cookieAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTelemetryList'
          description: ''
...
components:
  schemas:
    Alert:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        device:
          type: integer
        device_hardware_identifier:
          type: string
          readOnly: true
        alert_type:
          type: string
          maxLength: 64
        status:
          $ref: '#/components/schemas/StatusEnum'
        triggered_at:
          type: string
          format: date-time
          readOnly: true
        resolved_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        owner_id:
          type: integer
          readOnly: true
        owner_email:
          type: string
          format: email
          readOnly: true
        owner_phone:
          type: string
          readOnly: true
      required:
      - alert_type
      - device
      - device_hardware_identifier
      - id
      - owner_email
      - owner_id
      - owner_phone
      - resolved_at
      - triggered_at
    Device:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        hardware_identifier:
          type: string
          maxLength: 64
        device_name:
          type: string
          maxLength: 255
        latitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
          nullable: true
        longitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
          nullable: true
        status:
          type: string
          maxLength: 32
        registered_at:
          type: string
          format: date-time
          readOnly: true
        last_seen:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        online:
          type: string
          readOnly: true
        device_role:
          type: string
          readOnly: true
        master_id:
          type: integer
          readOnly: true
        owner_id:
          type: integer
          readOnly: true
        owner_email:
          type: string
          format: email
          readOnly: true
        owner_phone:
          type: string
          readOnly: true
      required:
      - device_role
      - hardware_identifier
      - id
      - last_seen
      - master_id
      - online
      - owner_email
      - owner_id
      - owner_phone
      - registered_at
    DeviceNode:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        hardware_identifier:
          type: string
          maxLength: 64
        device_name:
          type: string
          maxLength: 255
        latitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
          nullable: true
        longitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
          nullable: true
        status:
          type: string
          maxLength: 32
        registered_at:
          type: string
          format: date-time
          readOnly: true
        last_seen:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        online:
          type: string
          readOnly: true
        device_role:
          type: string
          readOnly: true
        master_id:
          type: integer
          readOnly: true
        owner_id:
          type: integer
          readOnly: true
        owner_email:
          type: string
          format: email
          readOnly: true
        owner_phone:
          type: string
          readOnly: true
      required:
      - device_role
      - hardware_identifier
      - id
      - last_seen
      - master_id
      - online
      - owner_email
      - owner_id
      - owner_phone
      - registered_at
    DeviceRegister:
      type: object
      properties:
        hardware_identifier:
          type: string
          description: Unique hardware identifier of the device
          maxLength: 64
        device_name:
          type: string
          description: Optional name
          maxLength: 255
        latitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
          description: Latitude in decimal degrees (-90..90)
        longitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
          description: Longitude in decimal degrees (-180..180)
        device_role:
          allOf:
          - $ref: '#/components/schemas/DeviceRoleEnum'
          default: master
          description: |-
            Role of the device: master (default) or slave

            * `master` - Master
            * `slave` - Slave
        master_id:
          type: integer
          nullable: true
          description: Required when device_role=slave. Must reference an existing
            master device (ID) that belongs to you (unless superadmin).
      required:
      - hardware_identifier
      - latitude
      - longitude
    DeviceRoleEnum:
      enum:
      - master
      - slave
      type: string
      description: |-
        * `master` - Master
        * `slave` - Slave
    DeviceTree:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        hardware_identifier:
          type: string
          maxLength: 64
        device_name:
          type: string
          maxLength: 255
        latitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
          nullable: true
        longitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
          nullable: true
        status:
          type: string
          maxLength: 32
        registered_at:
          type: string
          format: date-time
          readOnly: true
        last_seen:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        online:
          type: string
          readOnly: true
        device_role:
          type: string
          readOnly: true
        master_id:
          type: integer
          readOnly: true
        owner_id:
          type: integer
          readOnly: true
        owner_email:
          type: string
          format: email
          readOnly: true
        owner_phone:
          type: string
          readOnly: true
        slaves:
          type: array
          items:
            $ref: '#/components/schemas/DeviceNode'
          readOnly: true
      required:
      - device_role
      - hardware_identifier
      - id
      - last_seen
      - master_id
      - online
      - owner_email
      - owner_id
      - owner_phone
      - registered_at
      - slaves
...
    PaginatedAlertList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Alert'
...
    PaginatedDeviceTreeList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/DeviceTree'
    PaginatedTelemetryList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Telemetry'
...
    StatusEnum:
      enum:
      - open
      - resolved
      type: string
      description: |-
        * `open` - Open
        * `resolved` - Resolved
    Telemetry:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        device_id:
          type: integer
          writeOnly: true
        device:
          type: integer
          readOnly: true
        device_hardware_identifier:
          type: string
          readOnly: true
        smoke_level:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        device_status:
          type: string
          maxLength: 32
        timestamp:
          type: string
          format: date-time
        received_at:
          type: string
          format: date-time
          readOnly: true
        owner_id:
          type: integer
          readOnly: true
        owner_email:
          type: string
          format: email
          readOnly: true
        owner_phone:
          type: string
          readOnly: true
      required:
      - device
      - device_hardware_identifier
      - device_id
      - device_status
      - id
      - owner_email
      - owner_id
      - owner_phone
      - received_at
      - smoke_level
      - timestamp
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT